requirements.txt
aiofiles==23.2.1
httpx==0.23.3
Pillow==9.4.0
PyYAML==6.0
